Add State and Exit classes
State class acts as a wrapper for the named states, and also holds
the set of allowed exit paths.
Exit class is a basic holder of a name and an array of named guards,
that will be accessed through a Container object.
Also add the custom exception classes that the gem can raise.